Update .gitconfig to didable pager
It seems that in the newer version of git the default to use the pager
for every command was enabled, so I've updateded to config to use `cat`
as the default pager. `diff` has explicitly been configured to use
`less` so that it will still page like it did before.